Adding error reporting handlers with verbose opt
One reason I am writing this script is to provide Zabbix with a simple
up/down error code for testing LDAP availability. However, I still want
to be able to use it to read out LDAP searches in a verbose mode. Maybe
it doesn't make sense to have all of this in a config file, but for now
it's the right setup for my Zabbix use.